refactor(web): colocate LLM header resolution
